requirements.txt
selenium==4.12.0
beautifulreport==0.1.3
Jinja2==3.1.2
DBUtils==3.0.3
pymysql==1.1.0
pyyaml==6.0.1
loguru==0.7.1
webdriver-manager==4.0.0
XTestRunner==1.7.0
pylightxl==1.61.0
